pyproject.toml
[project]
name = "agi-consciousness"
version = "0.1.0"
description = "Processual consciousness model — multi-agent system with runtime states, hysteresis, and event loop"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.5.16",
    "openai>=1.60.0",
    "textual>=8.2.0",
    "structlog>=25.0.0",
    "pyyaml>=6.0",
    "rich>=14.0.0",
]

[project.scripts]
consciousness = "src.main:main"

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
